Adding ML Models of Waze Business Case
Adding the jupyter notebook and the executive summary of the Machine Learning Models project of the Waze Business Case. In this project, I generate a set of Random Forest (RF) and Extreme Gradient Boosting (XGBoost) models to explain the churn rate of a data extract similar to Waze's data (test dataset provided by the Google Advanced Data Analytics Certificate). In doing that, I applied a training/validation/test split methodology to select the champion model and provided key insights and next steps to improve the performance of the final model.